requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
PyMySQL>=1.1.0
Pillow>=10.0.0
pyserial>=3.5
numpy>=1.24.0
python-dotenv>=1.0.0 